pyproject.toml
[project]
name = "251022-libcst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "libcst>=1.8.5",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.12.0",
]
